Add learning curve data
